expose only builder with java function interface.
